requirements.txt
discord.py
qrcode
psutil
chat-exporter
configparser
